applications/space-data-collector/requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pika==1.3.2
pluggy==1.5.0
pytest==8.2.0
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
urllib3==2.2.1
Werkzeug==3.0.3
